feat: unifying into a module
