--- android/src/main/java/org/example/android/MainActivity.kt
@@ -1,25 +1,0 @@
-package org.example.android
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
-import org.example.shared.ui.App
-
-class MainActivity : ComponentActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        setContent {
-            App()
-        }
-    }
-}
-
-@Preview
-@Composable
-fun DefaultPreview() {
-    App()
-}
