+++ application/src/androidMain/kotlin/org/example/core/Plataform.kt
@@ -1,0 +1,5 @@
+package org.example.core
+
+actual fun getPlatform(): Platform {
+    return Platform.ANDROID
+}
