+++ application/src/androidMain/res/values/strings.xml
@@ -1,0 +1,3 @@
+<resources>
+    <string name="app_name">Android Example</string>
+</resources>
