+++ application/src/commonMain/kotlin/org/example/core/Platform.kt
@@ -1,0 +1,8 @@
+package org.example.core
+
+enum class Platform() {
+    DESKTOP,
+    ANDROID;
+}
+
+expect fun getPlatform(): Platform
