+++ application/src/commonMain/kotlin/org/example/ui/App.kt
@@ -1,0 +1,22 @@
+package org.example.ui
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import org.example.core.getPlatform
+
+@Composable
+fun App() = MaterialTheme {
+
+    Box(
+        contentAlignment = Alignment.Center,
+        modifier = Modifier.fillMaxSize()
+    ) {
+
+        Text("Hello, ${getPlatform()}!")
+    }
+}
