+++ application/src/desktopMain/kotlin/org/example/ui/main.kt
@@ -1,0 +1,33 @@
+package org.example.ui
+
+import androidx.compose.desktop.ui.tooling.preview.Preview
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.window.Window
+import androidx.compose.ui.window.WindowPosition
+import androidx.compose.ui.window.application
+import androidx.compose.ui.window.rememberWindowState
+
+fun main() = application {
+
+    Window(
+        onCloseRequest = ::exitApplication,
+        title = "Desktop Example",
+        state = rememberWindowState(
+            width = 300.dp,
+            height = 300.dp,
+            position = WindowPosition.Aligned(
+                Alignment.Center
+            )
+        )
+    ) {
+        App()
+    }
+}
+
+@Preview
+@Composable
+fun DefaultPreview() {
+    App()
+}
