+++ gradle/libs.versions.toml
@@ -3,15 +3,14 @@ agp = "8.2.2" # Kotlin Plugin 1.9.22 support APG 7.0 ~ 8.2
 compose = "1.6.3"
 compose-plugin = "1.6.1"
 kotlin = "1.9.22"
-androidx-activityCompose = "1.8.2"
-lifecycleViewmodelKtx = "2.7.0"
+androidx-activity = "1.8.2"
 
 [libraries]
-androidx-lifecycle-viewmodel-ktx = { module = "androidx.lifecycle:lifecycle-viewmodel-ktx", version.ref = "lifecycleViewmodelKtx" }
 compose-ui-tooling = { module = "androidx.compose.ui:ui-tooling", version.ref = "compose" }
 compose-ui-tooling-preview = { module = "androidx.compose.ui:ui-tooling-preview", version.ref = "compose" }
 
-androidx-activity-compose = { module = "androidx.activity:activity-compose", version.ref = "androidx-activityCompose" }
+androidx-activity = { module = "androidx.activity:activity", version.ref = "androidx-activity" }
+androidx-activity-compose = { module = "androidx.activity:activity-compose", version.ref = "androidx-activity" }
 
 [plugins]
 android-application = { id = "com.android.application", version.ref = "agp" }
