--- shared/src/commonMain/kotlin/org/example/shared/core/Platform.kt
@@ -1,8 +1,0 @@
-package org.example.shared.core
-
-enum class Platform() {
-    DESKTOP,
-    ANDROID;
-}
-
-expect fun getPlatform(): Platform
