--- shared/src/commonMain/kotlin/org/example/shared/ui/App.kt
@@ -1,22 +1,0 @@
-package org.example.shared.ui
-
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import org.example.shared.core.getPlatform
-
-@Composable
-fun App() = MaterialTheme {
-
-    Box(
-        contentAlignment = Alignment.Center,
-        modifier = Modifier.fillMaxSize()
-    ) {
-
-        Text("Hello, ${getPlatform()}!")
-    }
-}
