--- shared/src/desktopMain/kotlin/org/example/shared/core/Plataform.kt
@@ -1,5 +1,0 @@
-package org.example.shared.core
-
-actual fun getPlatform(): Platform {
-    return Platform.DESKTOP
-}
